fix: :bug: fixed issue in api init command
