+++ package.json
@@ -40,8 +40,8 @@
   "devDependencies": {
     "@eslint/js": "^9.2.0",
     "@release-it/conventional-changelog": "^8.0.1",
-    "@stylistic/eslint-plugin-js": "^2.0.0",
-    "@stylistic/eslint-plugin-ts": "^2.0.0",
+    "@stylistic/eslint-plugin-js": "^2.1.0",
+    "@stylistic/eslint-plugin-ts": "^2.1.0",
     "@tsconfig/node18": "^18.2.4",
     "@types/node": "^20.12.11",
     "@types/yargs": "^17.0.32",
