+++ src/commands/configure_cmds/init.ts
@@ -7,7 +7,7 @@ import {
   helpMessage,
 } from '../../utils/constants.js';
 import { handleConfigInit } from '../../handler/config/init/init.js';
-import { TargetProviders } from '../../types/configType.js';
+import { TargetProviders } from '../../types/enum-types.js';
 
 export const initCommand = {
   command: 'init',
@@ -27,7 +27,8 @@ export const initCommand = {
   handler: async (argv) => {
     try {
       await handleConfigInit(argv);
-      if (getTarget() !== TargetProviders.LOCAL) {
+      const target = getTarget();
+      if (target && target !== TargetProviders.LOCAL) {
         console.log(capabilitiesHelpMessage);
       }
       console.log(helpMessage);
